Enable copy/paste from PDF with special characters
From the PDF of the previous revision, copying text was limited as special characters were not copied properly. This commit changes the output encoding using the cmap package so that copying to plain text also includes special characters. In order for the cmap package to work, the font encoding had to be changed to T1.